refactor: Move CoinRepository to core/domain module
- Move CoinRepository, CoinInfoVO, CoinPriceVO from home/domain to core/domain
- Update package paths from home.domain to core.domain
- Add core/domain dependency to home/domain module
