+++ app/src/main/java/io/soma/cryptobook/di/RepositoryModule.kt
@@ -4,8 +4,8 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import io.soma.cryptobook.core.domain.repository.CoinRepository
 import io.soma.cryptobook.home.data.repository.CoinRepositoryImpl
-import io.soma.cryptobook.home.domain.repository.CoinRepository
 import javax.inject.Singleton
 
 @Module
