+++ core/domain/src/main/java/io/soma/cryptobook/core/domain/model/CoinInfoVO.kt
@@ -1,4 +1,4 @@
-package io.soma.cryptobook.home.domain.model
+package io.soma.cryptobook.core.domain.model
 
 data class CoinInfoVO(
     val symbol: String,
