+++ home/domain/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     id("java-library")
     alias(libs.plugins.jetbrains.kotlin.jvm)
+    alias(libs.plugins.cryptobook.hilt)
     alias(libs.plugins.cryptobook.spotless)
 }
 java {
@@ -14,6 +15,6 @@ kotlin {
 }
 
 dependencies {
-    implementation(libs.hilt.core)
+    api(projects.core.domain)
     implementation(libs.kotlinx.coroutines.core)
 }
